Optimize the Layout ViewBinding
